refactor(create-bananass): migrate to typescript
