+++ packages/bananass-utils-console/package.json
@@ -56,13 +56,9 @@
     }
   },
   "files": [
-    "src",
     "build",
     "LICENSE.md",
-    "README.md",
-    "!src/**/*.test.{js,ts}",
-    "!src/**/*.test-d.ts",
-    "!**/fixtures/**"
+    "README.md"
   ],
   "keywords": [
     "bananass",
