+++ packages/create-bananass/tsconfig.json
@@ -16,6 +16,7 @@
     "rootDir": "./src",
 
     /* Emit */
+    "emitDeclarationOnly": false,
     "outDir": "build"
   }
 }
