+++ packages/eslint-config-bananass/package.json
@@ -84,13 +84,9 @@
     }
   },
   "files": [
-    "src",
     "build",
     "LICENSE.md",
-    "README.md",
-    "!src/**/*.test.{js,ts}",
-    "!src/**/*.test-d.ts",
-    "!**/fixtures/**"
+    "README.md"
   ],
   "keywords": [
     "bananass",
